opt fix
Due to a two similar layer naming conventions between opt and llama, elif blocks were competing for meeting each other's qualifiers, leading llama to be unaffected but opt to attempt to merge as llama and throw an error that halts the operation. Revision clarifies attributes that qualify the elif block for relevant models.